@@ -1,24 +1,17 @@
-import {Component, OnInit} from '@angular/core';
-import {NavService} from "../../../../services/nav/nav.service";
+import { Component, ChangeDetectionStrategy, inject, OnInit } from '@angular/core';
+import { Title } from '@angular/platform-browser';
 import {I18nService} from "../../../../services/i18n/i18n.service";
-import {Title} from "@angular/platform-browser";
 
 @Component({
   selector: 'app-cgv',
   templateUrl: './cgv.component.html',
-  standalone: false
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CgvComponent implements OnInit {
-
-  constructor(private titleService: Title, private navService: NavService, public i18n: I18nService) {
-  }
+  public i18n = inject(I18nService);
+  private titleService = inject(Title);
 
   ngOnInit(): void {
-    window.scrollTo(0, 0);
-
-    this.navService.updateNav();
-
-    this.titleService.setTitle("Gaetan • " + this.i18n.text.title.cgv);
+    this.titleService.setTitle("Gaetan • " + this.i18n.text().title.cgv);
   }
-
 }
