@@ -4,7 +4,7 @@ import {RouterModule, Routes} from "@angular/router";
 import {AdminComponent} from "./admin.component";
 import {AuthComponent} from "./auth/auth.component";
 import {WhitelistComponent} from "./whitelist/whitelist.component";
-import {AuthGuard} from "../../../guard/auth.guard";
+import {AuthGuard} from "../../../guards/auth.guard";
 import {ProxiesComponent} from "./proxies/proxies.component";
 import {DetailserverComponent} from "./proxies/detailserver/detailserver.component";
 import {EmailComponent} from "./email/email.component";
