@@ -1,10 +1,10 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
 import { AdminService } from 'src/app/services/admin/admin.service';
 import { toast } from 'ngx-sonner';
 import { interval, Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { z } from 'zod';
-import { CommonModule, NgFor, NgIf } from '@angular/common';
+
 import { EmailsLoadingSkeletonsComponent } from '../../../include/skeletons/emails-loading-skeletons/emails-loading-skeletons.component';
 import { AddEmailSkeletonsComponent } from '../../../include/skeletons/add-email-skeletons/add-email-skeletons.component';
 import { TempladminComponent } from '../../../include/admin/templadmin/templadmin.component';
@@ -24,15 +24,12 @@ interface WordsResponse {
 @Component({
   selector: 'app-email',
   imports: [
-    CommonModule,
     FormsModule,
-    NgFor,
-    NgIf,
     TempladminComponent,
     EmailsLoadingSkeletonsComponent,
     AddEmailSkeletonsComponent,
     RemoveEmailSkeletonsComponent
-  ],
+],
   templateUrl: './email.component.html',
 })
 export class EmailComponent implements OnInit, OnDestroy {
@@ -50,7 +47,7 @@ export class EmailComponent implements OnInit, OnDestroy {
   secondWords: string[] = [];
   wordsLoaded: boolean = false;
 
-  constructor(private adminService: AdminService) {}
+  constructor(private adminService: AdminService, private cdr: ChangeDetectorRef) {}
 
   ngOnInit(): void {
     this.updateEmails().catch(err =>
@@ -121,7 +118,6 @@ export class EmailComponent implements OnInit, OnDestroy {
 
     this.isSubmitting = true;
     const loadingToast = this.toast.loading("Creating alias email...");
-    await new Promise(resolve => setTimeout(resolve, 1000));
 
     this.adminService.addEmail(this.emailName.trim(), this.description.trim()).subscribe({
       next: (response: any) => {
@@ -153,7 +149,6 @@ export class EmailComponent implements OnInit, OnDestroy {
 
     this.removingEmail = emailToRemove.nom;
     const loadingToast = this.toast.loading("Deleting alias email...");
-    await new Promise(resolve => setTimeout(resolve, 1000));
 
     this.adminService.removeEmail(emailToRemove.nom).subscribe({
       next: (response: any) => {
@@ -177,19 +172,21 @@ export class EmailComponent implements OnInit, OnDestroy {
     if (this.isLoadingEmails) return;
     this.isLoadingEmails = true;
     const loadingToast = this.toast.loading("Refreshing alias emails...");
-    await new Promise(resolve => setTimeout(resolve, 1000));
 
     this.adminService.getEmails().subscribe({
-      next: (response: Email[]) => {
+      next: async (response: Email[]) => {
+        await new Promise(resolve => setTimeout(resolve, 1000));
         this.emails = response || [];
+        this.isLoadingEmails = false;
+        this.cdr.detectChanges();
         this.toast.success("Alias emails refreshed", { id: loadingToast });
       },
-      error: (error) => {
+      error: async (error) => {
+        await new Promise(resolve => setTimeout(resolve, 1000));
+        this.isLoadingEmails = false;
+        this.cdr.detectChanges();
         this.toast.error("Error fetching alias emails");
         console.error("Error fetching alias emails:", error);
-      },
-      complete: () => {
-        this.isLoadingEmails = false;
       }
     });
   }
