@@ -4,7 +4,7 @@ import { Component } from '@angular/core';
   selector: 'app-refresh-proxies-whitelist',
   imports: [],
   template: `
-    <div class="h-[400px] w-full bg-gray-300 rounded-lg relative overflow-hidden before:absolute before:inset-0 before:-translate-x-full before:animate-[shimmer_2s_infinite] before:bg-gradient-to-r before:from-transparent before:via-white/60 before:to-transparent"></div>
+    <div class="h-[400px] w-full bg-white/10 backdrop-blur-md rounded-xl border border-white/20 relative overflow-hidden before:absolute before:inset-0 before:-translate-x-full before:animate-[shimmer_2s_infinite] before:bg-gradient-to-r before:from-transparent before:via-white/60 before:to-transparent"></div>
   `
 })
 export class RefreshProxiesWhitelistComponent {
