@@ -1,317 +1,349 @@
-import {EventEmitter, Injectable} from '@angular/core';
-import {LocalstorageService} from "../localstorage/localstorage.service";
-import {Router} from "@angular/router";
+import { Injectable, signal, computed, effect } from '@angular/core';
 
 @Injectable({
   providedIn: 'root'
 })
 export class I18nService {
-  isFrench: boolean = false;
-  text: any;
-  updateEvent: EventEmitter<any> = new EventEmitter();
+  private _isEnglish = signal<boolean>(this.getInitialLanguage());
 
-  constructor(private localStorage: LocalstorageService, private router: Router) {
-    const englishQueryParam: boolean = window.location.href.includes("/en");
-    const frenchQueryParam: boolean = window.location.href.includes("/fr");
-    const hasCache: boolean = (this.localStorage.get.getItem("english") === "true");
+  public readonly isEnglish = this._isEnglish.asReadonly();
+  public readonly text = computed(() => this.getTexts(this.isEnglish()));
 
-    if (englishQueryParam) {
-      this.localStorage.get.setItem("english", "true");
-      this.isFrench = true;
-      this.router.navigate([""]).then(() => console.log("Redirected to english version"));
-    } else if (frenchQueryParam) {
-      this.localStorage.get.setItem("english", "false");
-      this.isFrench = false;
-      this.router.navigate([""]).then(() => console.log("Redirected to french version"));
-    } else if (hasCache) {
-      this.isFrench = true;
-    }
-
-    this.update();
+  constructor() {
+    effect(() => {
+      try {
+        localStorage.setItem('isEnglish', JSON.stringify(this.isEnglish()));
+      } catch (e) {
+        console.error('Could not access localStorage.', e);
+      }
+    });
   }
 
+  private getInitialLanguage(): boolean {
+    // First, try to detect from URL if language prefix is present
+    if (typeof window !== 'undefined') {
+      const urlMatch = window.location.pathname.match(/^\/(fr|en)(\/|$)/);
+      if (urlMatch) {
+        return urlMatch[1] === 'en';
+      }
+    }
+    
+    // If no language prefix in URL, use localStorage or default to French
+    try {
+      const storedLang = localStorage.getItem('isEnglish');
+      return storedLang ? JSON.parse(storedLang) : false; // Default to French
+    } catch (e) {
+      console.error('Could not access localStorage.', e);
+      return false;
+    }
+  }
 
-  update(): void {
-    this.text = {
-      "lang": (this.isFrench ? "Langue changé en français." : "Language changed to english."),
+  public toggleLanguage(): void {
+    this._isEnglish.update(isEnglish => !isEnglish);
+  }
+
+  public setLanguage(lang: 'fr' | 'en'): void {
+    this._isEnglish.set(lang === 'en');
+  }
+
+  private getTexts(isEnglish: boolean) {
+    return {
+      "lang": (isEnglish ? "Language changed to english." : "Langue changé en français."),
       "title": {
-        "about": (this.isFrench ? "About Me" : "A Propos"),
-        "achievements": (this.isFrench ? "Achievements" : "Réalisations"),
-        "cgv": (this.isFrench ? "TOS/CGV" : "CGU/CGV"),
-        "legal": (this.isFrench ? "Legal mention" : "Mention légales"),
+        "about": (isEnglish ? "About Me" : "A Propos"),
+        "achievements": (isEnglish ? "Achievements" : "Réalisations"),
+        "cgv": (isEnglish ? "TOS/CGV" : "CGU/CGV"),
+        "legal": (isEnglish ? "Legal mention" : "Mention légales"),
       },
-      "devWithLove": (this.isFrench ?
-          "Developed and designed with ❤️ by Gaëtan Faucher" :
-          "Développé et conçu avec ❤️ par Gaëtan Faucher"
-      ),
-      "learnMoreButton": (this.isFrench ? "Learn more" : "En savoir plus"),
-      "learnMoreButtonAchievements": (this.isFrench ? "See more" : "Voir plus de réalisations"),
-      "desc": (this.isFrench ?
-          "I am a young full-stack developer and DevOps ! I am very interested in new technologies, development and computing." :
+      "learnMoreButton": (isEnglish ? "Learn more" : "En savoir plus"),
+      "desc": (isEnglish ?
+          "I am a young full-stack developer and DevOps! I am very interested in new technologies, development and computing." :
           "Je suis un jeune développeur full-stack et DevOps ! Je m'intéresse beaucoup aux nouvelles technologies, au développement et à l'informatique."
       ),
-
-      //include
       "navbar": {
-        "home": (this.isFrench ? "Home" : "Accueil"),
-        "about": (this.isFrench ? "About Me" : "A Propos"),
-        "achievements": (this.isFrench ? "Achievements" : "Réalisations"),
+        "home": (isEnglish ? "Home" : "Accueil"),
+        "about": (isEnglish ? "About Me" : "A Propos"),
+        "achievements": (isEnglish ? "Achievements" : "Réalisations"),
       },
       "footer": {
-        "legal": (this.isFrench ? "Legal mention" : "Mention légales"),
-        "cgu": (this.isFrench ? "TOS/CGV" : "CGU/CGV"),
+        "legal": (isEnglish ? "Legal mention" : "Mention légales"),
+        "cgv": (isEnglish ? "TOS/CGV" : "CGU/CGV"),
+        "devWithLove": (isEnglish ? "Developed and designed with ❤️ by Gaëtan Faucher" : "Développé et conçu avec ❤️ par Gaëtan Faucher"),
+      },
+      "home": {
+        "welcome": (isEnglish ? "Hello, I'm Gaëtan" : "Bonjour, je suis Gaëtan"),
+        "contactButton": (isEnglish ? "Contact me" : "Me contacter"),
+      },
+      "contact": {
+        "title": (isEnglish ? "Contact me" : "Me contacter"),
+        "nameLabel": (isEnglish ? "Name" : "Nom"),
+        "emailLabel": "Email",
+        "messageLabel": "Message",
+        "sendButton": (isEnglish ? "Send Message" : "Envoyer le message"),
+        "sendingButton": (isEnglish ? "Sending..." : "Envoi en cours..."),
+        "validation": {
+          "nameRequired": (isEnglish ? "Your name is required." : "Votre nom est requis."),
+          "nameMinLength": (isEnglish ? "Your name must be at least 3 characters long." : "Votre nom doit comporter au moins 3 caractères."),
+          "emailRequired": (isEnglish ? "Your email is required." : "Votre e-mail est requis."),
+          "emailInvalid": (isEnglish ? "Please enter a valid email address." : "Veuillez entrer une adresse e-mail valide."),
+          "messageRequired": (isEnglish ? "A message is required." : "Un message est requis."),
+          "captchaInvalid": (isEnglish ? "Verification failed. Please confirm you're not a robot." : "Vérification échouée. Merci de confirmer que vous n'êtes pas un robot."),
+        },
+        "form": {
+          "nameInvalid": (isEnglish ? "Your name must be at least 3 characters long." : "Votre nom doit comporter au moins 3 caractères."),
+          "mailInvalid": (isEnglish ? "Please enter a valid email address." : "Veuillez entrer une adresse e-mail valide."),
+          "messageInvalid": (isEnglish ? "A message is required." : "Un message est requis."),
+          "captchaInvalid": (isEnglish ? "Verification failed. Please confirm you're not a robot." : "Vérification échouée. Merci de confirmer que vous n'êtes pas un robot."),
+          "limit": (isEnglish ? "Please wait a moment before sending another message." : "Veuillez patienter quelques instants avant d'envoyer un nouveau message."),
+          "loading": (isEnglish ? "Sending..." : "Envoi en cours..."),
+          "success": (isEnglish ? "Your message has been sent successfully! I'll get back to you soon." : "Votre message a été envoyé avec succès ! Je vous répondrai bientôt."),
+          "server": (isEnglish ? "An error occurred. Please try again later." : "Une erreur est survenue. Veuillez réessayer plus tard."),
+        },
+        "limitMessage": (isEnglish ? "Please wait a moment before sending another message." : "Veuillez patienter quelques instants avant d'envoyer un nouveau message."),
+        "successMessage": (isEnglish ? "Your message has been sent successfully! I'll get back to you soon." : "Votre message a été envoyé avec succès ! Je vous répondrai bientôt."),
+        "errorMessage": (isEnglish ? "An error occurred. Please try again later." : "Une erreur est survenue. Veuillez réessayer plus tard."),
+      },
+      "about": {
+        "title": (isEnglish ? "About me" : "A propos de moi"),
+        "role": (isEnglish ? "Hello, I am Gaëtan Faucher." : "Bonjour, je suis Gaëtan Faucher."),
+        "cvButton": (isEnglish ? "Download my CV" : "Télécharger mon CV"),
+        "cvNotAvailable": (isEnglish ? "CV is not available at the moment." : "Le CV n'est pas disponible pour le moment."),
+      },
+      "achievements": {
+        "title": (isEnglish ? "My Achievements" : "Mes Réalisations"),
+        "description": (isEnglish ? "Here's a list of my accomplishments that I am allowed to showcase publicly. Please note that some of my accomplishments are not included on this list due to non-disclosure agreements in effect." : "Voici une liste de mes réalisations dont j'ai l'autorisation de partager publiquement. Veuillez noter que certaines de mes réalisations ne figurent pas sur cette liste en raison de contrats de confidentialité en vigueur."),
+        "list": [
+          {
+            "name": "TMConseils",
+            "description": (!isEnglish ? "Application web pour une entreprise d'accompagnement stratégique sur mesure." : "Web application for a customized strategic coaching company."),
+            "image": "assets/img/achievements/tmconseils.webp",
+            "redirect": "https://tm-conseils.net/",
+            "categories": ["DEV"]
+          },
+          {
+            "name": "LookAway",
+            "description": (!isEnglish ? "Application macOS pour protéger vos yeux en vous invitant à faire des pauses régulières." : "A macOS menu-bar application to protect your eyes by prompting regular breaks."),
+            "image": "assets/img/achievements/lookaway.webp",
+            "redirect": "https://github.com/GaetanOff/LookAway",
+            "categories": ["DEV"]
+          },
+          {
+            "name": "Content Delivery Network",
+            "description": (isEnglish ? "My content distribution network to be able to have a fast and reliable distribution." : "Mon réseau de distribution de contenu pour pouvoir avoir une diffusion rapides et fiables."),
+            "image": "assets/img/achievements/cdn.webp",
+            "redirect": "https://cdn.gaetandev.fr/",
+            "categories": ["SYS", "DO"]
+          },
+          {
+            "name": "TCPing",
+            "description": (!isEnglish ? "Outil léger de test de connexion TCP." : "A lightweight TCP connection testing tool."),
+            "image": "assets/img/achievements/tcping.webp",
+            "redirect": "https://github.com/GaetanOff/TCPing",
+            "categories": ["SYS", "DO"]
+          },
+          {
+            "name": "PixelmonGo",
+            "description": (isEnglish ? "Management of the infrastructure of the entire server." : "Gestion de l'infrastructure de l'ensemble du serveur."),
+            "image": "assets/img/achievements/pixelmongo.webp",
+            "redirect": "https://pixelmongo.fr/",
+            "categories": ["SYS"]
+          },
+          {
+            "name": "FirstSky",
+            "description": (isEnglish ? "Management of the infrastructure of the entire server." : "Gestion de l'infrastructure de l'ensemble du serveur."),
+            "image": "assets/img/achievements/firstsky.webp",
+            "redirect": "https://firstsky.net/",
+            "categories": ["SYS"]
+          },
+          {
+            "name": "Hôtel Neptune",
+            "description": (isEnglish ? "School project for creating a web application for a fictitious hotel." : "Projet d'école de création d'une application web pour un hôtel fictif."),
+            "image": "assets/img/achievements/neptune.webp",
+            "redirect": "https://github.com/Projet-Hotel-Neptune-GLJQR/Application",
+            "categories": ["DEV", "DO"]
+          },
+          {
+            "name": "FoxFood",
+            "description": (isEnglish ? "School project for creating a fictitious company." : "Projet d'école de création d'une entreprise fictive."),
+            "image": "assets/img/achievements/foxfood.svg",
+            "redirect": "https://github.com/Fox-Food",
+            "categories": ["DEV", "DO"]
+          },
+          {
+            "name": "ImmoCT",
+            "description": (isEnglish ? "Web application for an independent real estate agent that allows them to showcase their listings." : "Application web pour un agent immobilier indépendant qui permet de mettre en avant ses offres."),
+            "image": "assets/img/achievements/immoct.webp",
+            "redirect": "https://www.immoct.fr/",
+            "categories": ["DEV", "DO"]
+          },
+          {
+            "name": "CT-Energetique",
+            "description": (isEnglish ? "Web application for a Chinese medicine company." : "Application web pour une entreprise de médécine chinoise."),
+            "image": "assets/img/achievements/ctenergetique.webp",
+            "redirect": "https://www.ct-energetique.fr/",
+            "categories": ["DEV", "DO"]
+          },
+          {
+            "name": "HeavyCloud",
+            "description": (isEnglish ? "Creation of a hosting provider offering VPS as well as Anti-DDoS solutions." : "Création d'un hébergeur proposant des VPS ainsi que des solutions Anti-DDoS."),
+            "image": "assets/img/achievements/heavycloud.webp",
+            "redirect": "",
+            "categories": ["DEV"]
+          },
+          {
+            "name": "ChloeCM",
+            "description": (isEnglish ? "Web application to showcase the journey of a Master's degree student." : "Application web pour présenter le parcours d'une étudiante en Master."),
+            "image": "assets/img/achievements/chloecm.webp",
+            "redirect": "https://www.chloecm.fr/",
+            "categories": ["DEV"]
+          },
+          {
+            "name": "VitaPot",
+            "description": (isEnglish ? "Management of the development of an practice minecraft server." : "Gestion du développement d'un serveur minecraft practice."),
+            "image": "assets/img/achievements/VitaPot.webp",
+            "redirect": "https://twitter.com/vitapot",
+            "categories": ["DEV", "SYS"]
+          },
+          {
+            "name": "PvPIng",
+            "description": (isEnglish ? "Development of an practice minecraft server." : "Développement d'un serveur minecraft practice."),
+            "image": "assets/img/achievements/pvping.webp",
+            "redirect": "https://twitter.com/ServeurPvPing",
+            "categories": ["DEV", "SYS"]
+          },
+          {
+            "name": "RealMC",
+            "description": (isEnglish ? "Creation/Management/Development of a practice minecraft server." : "Création/Gestion/Développement d'un serveur minecraft practice."),
+            "image": "assets/img/achievements/realmc.webp",
+            "redirect": "https://youtu.be/SD1547llpvI",
+            "categories": ["DEV", "SYS"]
+          },
+          {
+            "name": "LiptonPvP",
+            "description": (isEnglish ? "Creation/Management/Development of a practice minecraft server." : "Création/Gestion/Développement d'un serveur minecraft practice."),
+            "image": "assets/img/achievements/liptonpvp.webp",
+            "redirect": "https://youtu.be/l5aYEVQmop8",
+            "categories": ["DEV", "SYS"]
+          }
+        ]
       },
       "services": {
-        "title": (this.isFrench ? "What I offer" : "Ce que je propose"),
-        "devis": (this.isFrench ? "Quotation" : "Sur devis"),
-        "contactButton": (this.isFrench ? "Contact me" : "Contactez-moi"),
-        "perDay": "/" + (this.isFrench ? "day" : "jour"),
-        "devTitle": (this.isFrench ? "Development" : "Développement"),
+        "title": (isEnglish ? "What I offer" : "Ce que je propose"),
+        "devis": (isEnglish ? "Quotation" : "Sur devis"),
+        "contactButton": (isEnglish ? "Contact me" : "Contactez-moi"),
+        "perDay": "/" + (isEnglish ? "day" : "jour"),
+        "devTitle": (isEnglish ? "Development" : "Développement"),
         "dev": [
-          (this.isFrench ? "Custom development" : "Développement sur mesure"),
-          (this.isFrench ? "Multi-platform" : "Multi plateforme"),
-          (this.isFrench ? "Advice" : "Conseil"),
+          (isEnglish ? "Custom development" : "Développement sur mesure"),
+          (isEnglish ? "Multi-platform" : "Multi plateforme"),
+          (isEnglish ? "Advice" : "Conseil"),
         ],
+        "devopsTitle": "DevOps",
         "devops": [
-          (this.isFrench ? "CI/CD Pipelines" : "Pipelines CI/CD"),
-          (this.isFrench ? "Containerization (Docker/K8s)" : "Conteneurisation (Docker/K8s)"),
-          (this.isFrench ? "Infrastructure as Code" : "Infrastructure as Code"),
+          (isEnglish ? "CI/CD Pipelines" : "Pipelines CI/CD"),
+          (isEnglish ? "Containerization (Docker/K8s)" : "Conteneurisation (Docker/K8s)"),
+          (isEnglish ? "Infrastructure as Code" : "Infrastructure as Code"),
         ],
-        "iaTitle": (this.isFrench ? "AI Solutions" : "Solutions IA"),
+        "iaTitle": (isEnglish ? "AI Solutions" : "Solutions IA"),
         "ia": [
-          (this.isFrench ? "Model integration" : "Intégration de modèles"),
-          (this.isFrench ? "Workflow automation" : "Automatisation de workflows"),
-          (this.isFrench ? "AI audit & strategy" : "Audit & stratégie IA"),
+          (isEnglish ? "Model integration" : "Intégration de modèles"),
+          (isEnglish ? "Workflow automation" : "Automatisation de workflows"),
+          (isEnglish ? "AI audit & strategy" : "Audit & stratégie IA"),
         ]
       },
       "cv": {
         "competences": {
-          "title": (this.isFrench ? "Main technologies used" : "Principales technologies utilisées"),
+          "title": (isEnglish ? "Main technologies used" : "Principales technologies utilisées"),
         },
         "certif": {
-          "title": (this.isFrench ? "Certificates" : "Certifications"),
+          "title": (isEnglish ? "Certificates" : "Certifications"),
         },
         "experience": {
-          "title": (this.isFrench ? "Experiences" : "Expériences"),
+          "title": (isEnglish ? "Experiences" : "Expériences"),
         },
         "languages": {
-          "title": (this.isFrench ? "Languages" : "Langues"),
+          "title": (isEnglish ? "Languages" : "Langues"),
           "langList": [
-            (this.isFrench ? "French" : "Français"),
-            (this.isFrench ? "English" : "Anglais"),
+            (isEnglish ? "French" : "Français"),
+            (isEnglish ? "English" : "Anglais"),
           ]
-        }
-      },
-
-      //base
-      "home": {
-        "welcome": (this.isFrench ? "Hello, I'm Gaëtan" : "Bonjour, je suis Gaëtan"),
-        "contactButton": (this.isFrench ? "Contact me" : "Me contacter"),
-      },
-      "about": {
-        "title": (this.isFrench ? "About me" : "A propos de moi"),
-        "role": (this.isFrench ? "Hello, I am Gaëtan Faucher." : "Bonjour, je suis Gaëtan Faucher."),
-        "cvButton": (this.isFrench ? "Download my CV" : "Télécharger mon CV"),
-        "cvNotAvailable": (this.isFrench ? "This feature is not yet available" : "Cette fonctionnalité n'est pas encore disponible"),
-      },
-      "achievements": {
-        "title": (this.isFrench ? "My achievements" : "Mes réalisations"),
-        "description": (this.isFrench ? "Here's a list of my accomplishments that I am allowed to showcase publicly. Please note that some of my accomplishments are not included on this list due to non-disclosure agreements in effect." : "Voici une liste de mes réalisations dont j'ai l'autorisation de partager publiquement. Veuillez noter que certaines de mes réalisations ne figurent pas sur cette liste en raison de contrats de confidentialité en vigueur."),
-        "chloecm": (this.isFrench ? "Web application to showcase the journey of a Master's degree student." : "Application web pour présenter le parcours d'une étudiante en Master."),
-        "liptonpvp": (this.isFrench ? "Creation/Management/Developement of an practice minecraft server." : "Création/Gestion/Développement d'un serveur minecraft practice."),
-        "pvping": (this.isFrench ? "Development of an practice minecraft server." : "Développement d'un serveur minecraft practice."),
-        "vitapot": (this.isFrench ? "Management of the development of an practice minecraft server." : "Gestion du développement d'un serveur minecraft practice."),
-        "cdn": (this.isFrench ? "My content distribution network to be able to have a fast and reliable distribution." : "Mon réseau de distribution de contenu pour pouvoir avoir une diffusion rapides et fiables."),
-        "immoct": (this.isFrench ? "Web application for an independent real estate agent that allows them to showcase their listings." : "Application web pour un agent immobilier indépendant qui permet de mettre en avant ses offres."),
-        "ctenergetique": (this.isFrench ? "Web application for a Chinese medicine company." : "Application web pour une entreprise de médécine chinoise."),
-        "firstsky": (this.isFrench ? "Management of the infrastructure of the entire server." : "Gestion de l'infrastructure de l'ensemble du serveur."),
-        "pizzaoplomo": (this.isFrench ? "School project for creating a fictitious company." : "Projet d'école de création d'une entreprise fictive."),
-        "heavycloud": (this.isFrench ? "Creation of a hosting provider offering VPS as well as Anti-DDoS solutions." : "Création d'un hébergeur proposant des VPS ainsi que des solutions Anti-DDoS."),
-        "neptune": (this.isFrench ? "School project for creating a web application for a fictitious hotel." : "Projet d'école de création d'une application web pour un hôtel fictif."),
-        "tmconseils": (!this.isFrench
-          ? "Application web pour une entreprise d'accompagnement stratégique sur mesure."
-          : "Web application for a customized strategic coaching company."),
-        "lookaway": (!this.isFrench
-          ? "Application macOS pour protéger vos yeux en vous invitant à faire des pauses régulières."
-          : "A macOS menu-bar application to protect your eyes by prompting regular breaks."),
-        "tcping": (!this.isFrench
-          ? "Outil léger de test de connexion TCP."
-          : "A lightweight TCP connection testing tool.")
-      },
-      "shield": {
-        "title": (this.isFrench ? "Shield GaetanDev" : "Bouclier GaetanDev"),
-        "servers": (this.isFrench ? "Server currently protected" : "Serveur actuellement protégé"),
-        "protected": (this.isFrench ? "Attack received successfully mitigated" : "Attaque reçue atténuée avec succès"),
-        "total": (this.isFrench ? "Number of people connected to the infrastructure" : "Nombre de personnes connectées sur l'infrastructure"),
+        },
+        "today": (isEnglish ? "Today" : "Aujourd'hui"),
+        "unknownDuration": (isEnglish ? "Unknown duration" : "Durée inconnue"),
+        "lessThanMonth": (isEnglish ? "Less than a month" : "Moins d'un mois"),
+        "year": (isEnglish ? "year" : "an"),
+        "years": (isEnglish ? "years" : "ans"),
+        "month": (isEnglish ? "month" : "mois"),
+        "months": (isEnglish ? "months" : "mois"),
+        "and": (isEnglish ? "and" : "et"),
       },
       "error": {
         "notfound": {
-          "title": (this.isFrench ? "Page not found" : "Page introuvable"),
-          "description": (this.isFrench ? "The page you are looking for does not exist." : "La page que vous cherchez n'existe pas."),
-          "moved": (this.isFrench ? "It might have been moved or deleted." : "Elle pourrait avoir fait l'objet d'un transfert ou d'une suppression."),
-          "button": (this.isFrench ? "Go to the home page" : "Retourner à la page d'accueil")
-        }
-      },
-      "contact": {
-        "title": (this.isFrench ? "Contact me" : "Me contacter"),
-        "name": (this.isFrench ? "Name" : "Prénom"),
-        "sendButton": (this.isFrench ? "Send my message" : "Envoyer mon message"),
-        "form": {
-          "invalid": (this.isFrench ? "Please check the form and fix the errors." : "Veuillez vérifier le formulaire et corriger les erreurs."),
-          "server": (this.isFrench ? "The server is temporarily unavailable. Please try again later." : "Le serveur est temporairement indisponible. Veuillez réessayer plus tard."),
-          "nameInvalid": (this.isFrench ? "Please enter a valid first name." : "Merci de renseigner un prénom valide."),
-          "mailInvalid": (this.isFrench ? "Please enter a valid email address." : "Merci de renseigner une adresse e-mail valide."),
-          "messageInvalid": (this.isFrench ? "Your message seems empty or invalid." : "Votre message semble vide ou invalide."),
-          "captchaInvalid": (this.isFrench ? "Verification failed. Please confirm you're not a robot." : "Vérification échouée. Merci de confirmer que vous n'êtes pas un robot."),
-          "success": (this.isFrench ? "Your message has been successfully sent." : "Votre message a été envoyé avec succès."),
-          "loading": (this.isFrench ? "Sending, please wait..." : "Envoi en cours, merci de patienter..."),
-          "limit": (this.isFrench ? "Please wait a moment before sending another message." : "Veuillez patienter quelques instants avant d'envoyer un nouveau message.")
+          "title": (isEnglish ? "Page not found" : "Page introuvable"),
+          "description": (isEnglish ? "The page you are looking for does not exist." : "La page que vous cherchez n'existe pas."),
+          "moved": (isEnglish ? "It might have been moved or deleted." : "Elle pourrait avoir fait l'objet d'un transfert ou d'une suppression."),
+          "button": (isEnglish ? "Go to the home page" : "Retourner à la page d'accueil")
         }
       },
       "legal": {
-        "declaration": (this.isFrench ? "Statement" : "Déclaration"),
+        "declaration": (isEnglish ? "Statement" : "Déclaration"),
         "declarationText": [
-          (this.isFrench ?
-              "GaetanDev.fr edited by: FAUCHER Gaëtan Individual entrepreneur" :
-              "GaetanDev.fr édité par: FAUCHER Gaëtan Entrepreneur individuel"
-          ),
+          (isEnglish ? "GaetanDev.fr edited by: FAUCHER Gaëtan Individual entrepreneur" : "GaetanDev.fr édité par: FAUCHER Gaëtan Entrepreneur individuel"),
           "Siret: 91028598000017, CFE URSSAF 34",
-          (this.isFrench ? "Adress" : "Adresse") + ": 509 RUE DE BUGAREL 34070 MONTPELLIER",
-          (this.isFrench ? "Director of publication" : "Directeur de la publication") + ": Gaëtan Faucher - contact@gaetandev.fr",
+          (isEnglish ? "Adress" : "Adresse") + ": 509 RUE DE BUGAREL 34070 MONTPELLIER",
+          (isEnglish ? "Director of publication" : "Directeur de la publication") + ": Gaëtan Faucher - contact@gaetandev.fr",
         ],
-        "host": (this.isFrench ? "Host" : "Hébergeur"),
-        "hostText": [
-          this.isFrench ?
-            "The gaetandev.fr site is hosted on the servers of Amazon Web Services Inc, PO Box 81226, Seattle, WA 981808-1226 – USA. https://aws.amazon.com/en/compliance/eu-data-protection/.\n" :
-            "Le site gaetandev.fr est hébergé sur les serveurs de Amazon Web Services Inc, PO Box 81226, Seattle, WA 981808-1226 – USA. https://aws.amazon.com/fr/compliance/eu-data-protection/.",
-        ],
-        "reproductionText":
-          this.isFrench ?
-            "This entire site is subject to international legislation on copyright and intellectual property." :
-            "L'ensemble de ce site relève des législations internationales sur le droit d'auteur et la propriété intellectuelle.",
-        "copyrightText": this.isFrench ? "All reproduction rights are reserved." : "Tous les droits de reproduction sont réservés.",
-        "preamble": (this.isFrench ? "Preamble" : "Préambule"),
+        "host": (isEnglish ? "Host" : "Hébergeur"),
+        "hostText": (isEnglish ? "The gaetandev.fr site is hosted on the servers of Amazon Web Services Inc, PO Box 81226, Seattle, WA 981808-1226 – USA. https://aws.amazon.com/en/compliance/eu-data-protection/." : "Le site gaetandev.fr est hébergé sur les serveurs de Amazon Web Services Inc, PO Box 81226, Seattle, WA 981808-1226 – USA. https://aws.amazon.com/fr/compliance/eu-data-protection/."),
+        "reproductionTitle": (isEnglish ? "Reproduction" : "Reproduction"),
+        "reproductionText": (isEnglish ? "This entire site is subject to international legislation on copyright and intellectual property." : "L'ensemble de ce site relève des législations internationales sur le droit d'auteur et la propriété intellectuelle."),
+        "copyrightText": (isEnglish ? "All reproduction rights are reserved." : "Tous les droits de reproduction sont réservés."),
+        "preamble": (isEnglish ? "Preamble" : "Préambule"),
         "preambleText": [
-          (this.isFrench ?
-              "Access to the gaetandev.fr site is free, with no obligation to purchase." :
-              "L'accès au site gaetandev.fr est gratuit, sans obligation d'achat."
-          ),
-          (this.isFrench ?
-              "The gaetandev.fr site reserves the right to modify" :
-              "Le site gaetandev.fr se réserve le droit de modifier"
-          ),
-          (this.isFrench ?
-              "without notice or announcement of these general conditions of sale which must" :
-              "sans préavis ou annonce les présentes conditions générales de vente qui devront"
-          ),
-          (this.isFrench ?
-              "be consulted before any purchase. The user must be of legal age or" :
-              "être consultées avant tout achat. L'utilisateur doit être majeur ou"
-          ),
-          (this.isFrench ?
-              "have authorization from their manager before making a purchase." :
-              "disposer d'une autorisation de son responsable avant d'effectuer un achat."
-          ),
+          (isEnglish ? "Access to the gaetandev.fr site is free, with no obligation to purchase." : "L'accès au site gaetandev.fr est gratuit, sans obligation d'achat."),
+          (isEnglish ? "The gaetandev.fr site reserves the right to modify" : "Le site gaetandev.fr se réserve le droit de modifier"),
+          (isEnglish ? "without notice or announcement of these general conditions of sale which must" : "sans préavis ou annonce les présentes conditions générales de vente qui devront"),
+          (isEnglish ? "be consulted before any purchase. The user must be of legal age or" : "être consultées avant tout achat. L'utilisateur doit être majeur ou"),
+          (isEnglish ? "have authorization from their manager before making a purchase." : "disposer d'une autorisation de son responsable avant d'effectuer un achat."),
         ],
-        "definitions": (this.isFrench ? "Definitions" : "Définitions"),
+        "definitions": (isEnglish ? "Definitions" : "Définitions"),
         "definitionsText": [
-          (this.isFrench ?
-              "The services belonging to gaetandev.fr constitute all the services available at the URL https://gaetandev.fr/." :
-              "Les services appartenant à gaetandev.fr constituent l'ensemble des services disponibles à l'URL https://gaetandev.fr/."
-          )
+          (isEnglish ? "The services belonging to gaetandev.fr constitute all the services available at the URL https://gaetandev.fr/." : "Les services appartenant à gaetandev.fr constituent l'ensemble des services disponibles à l'URL https://gaetandev.fr/.")
         ],
-        "cgu": (this.isFrench ? "Terms of use" : "Conditions d'utilisation"),
+        "cgu": (isEnglish ? "Terms of use" : "Conditions d'utilisation"),
         "cguText": [
-          (this.isFrench ?
-              "All items offered on the gaetandev.fr website remain the exclusive property of" :
-              "Tous les articles proposés sur le site gaetandev.fr restent la propriété exclusive de"
-          ),
-          (this.isFrench ?
-              "gaetandev.fr. As such, no one has the right to sell or resell the content belonging to" :
-              "gaetandev.fr. À ce titre personne n'a le droit de vendre ou revendre le contenu appartenant à"
-          ),
-          (this.isFrench ?
-              "gaetandev.fr without permission. Failure to comply with the conditions of use or sale may lead to" :
-              "gaetandev.fr sans autorisation. Un non-respect des conditions d'utilisation ou de vente pourra conduire à des"
-          ),
-          (this.isFrench ?
-              "proceedings and a ban on the use of all gaetandev.fr services." :
-              "poursuites et une interdiction d'utilisation de tous les services de gaetandev.fr."
-          ),
+          (isEnglish ? "All items offered on the gaetandev.fr website remain the exclusive property of" : "Tous les articles proposés sur le site gaetandev.fr restent la propriété exclusive de"),
+          (isEnglish ? "gaetandev.fr. As such, no one has the right to sell or resell the content belonging to" : "gaetandev.fr. À ce titre personne n'a le droit de vendre ou revendre le contenu appartenant à"),
+          (isEnglish ? "gaetandev.fr without permission. Failure to comply with the conditions of use or sale may lead to" : "gaetandev.fr sans autorisation. Un non-respect des conditions d'utilisation ou de vente pourra conduire à des"),
+          (isEnglish ? "proceedings and a ban on the use of all gaetandev.fr services." : "poursuites et une interdiction d'utilisation de tous les services de gaetandev.fr."),
         ],
-
-        "command": (this.isFrench ? "Order process" : "Processus de commande"),
+        "command": (isEnglish ? "Order process" : "Processus de commande"),
         "commandText": [
-          (this.isFrench ?
-              "Any validated order constitutes an irrevocable acceptance of the general conditions of sale and the" :
-              "Toute commande validée constitue une acception irrévocable des conditions générales de vente et du"
-          ),
-          (this.isFrench ?
-              "payment and validate your transaction. The data recorded by gaetandev.fr constitute the proof" :
-              "paiement et validera votre transaction. Les données enregistrée par gaetandev.fr constituent la preuve"
-          ),
-          (this.isFrench ?
-              "of all transactions made by gaetandev.fr and its customers. The data recorded by" :
-              "de l'ensemble des transactions passées par gaetandev.fr et ses clients. Les données enregistrées par"
-          ),
-          (this.isFrench ?
-              "the payment system constitutes proof of financial transactions." :
-              "le système de paiement constitue la preuve des transactions financières."
-          ),
+          (isEnglish ? "Any validated order constitutes an irrevocable acceptance of the general conditions of sale and the" : "Toute commande validée constitue une acception irrévocable des conditions générales de vente et du"),
+          (isEnglish ? "payment and validate your transaction. The data recorded by gaetandev.fr constitute the proof" : "paiement et validera votre transaction. Les données enregistrée par gaetandev.fr constituent la preuve"),
+          (isEnglish ? "of all transactions made by gaetandev.fr and its customers. The data recorded by" : "de l'ensemble des transactions passées par gaetandev.fr et ses clients. Les données enregistrées par"),
+          (isEnglish ? "the payment system constitutes proof of financial transactions." : "le système de paiement constitue la preuve des transactions financières."),
         ],
-
-        "payment": (this.isFrench ? "Payment" : "Paiement"),
+        "payment": (isEnglish ? "Payment" : "Paiement"),
         "paymentText": [
-          (this.isFrench ?
-              "Payment is made online through Paypal." :
-              "Le paiement s'effectue en ligne par le biais de Paypal."
-          ),
-          (this.isFrench ?
-              "Personal information and credit card numbers never circulate"
-              : "Les informations personnelles ainsi que les numéros de carte bancaire ne circulent jamais"
-          ),
-          (this.isFrench ?
-              "in plain text between the browser and the server, the data is encrypted and transmitted by the SSL protocol." :
-              "en clair entre le navigateur et le serveur, les données sont chiffrées et transmises par le protocole SSL."
-          ),
+          (isEnglish ? "Payment is made online through Paypal." : "Le paiement s'effectue en ligne par le biais de Paypal."),
+          (isEnglish ? "Personal information and credit card numbers never circulate" : "Les informations personnelles ainsi que les numéros de carte bancaire ne circulent jamais"),
+          (isEnglish ? "in plain text between the browser and the server, the data is encrypted and transmitted by the SSL protocol." : "en clair entre le navigateur et le serveur, les données sont chiffrées et transmises par le protocole SSL."),
         ],
-
-        "retractation": (this.isFrench ? "Right to retract" : "Droit de rétractation"),
+        "retractation": (isEnglish ? "Right to retract" : "Droit de rétractation"),
         "retractationText": [
-          (this.isFrench ?
-              "Notwithstanding article L. 121-20-2 of the French Consumer Code," :
-              "Par dérogation à l'article L. 121-20-2 du Code Français de la consommation,"
-          ),
-          (this.isFrench ?
-            "the customer no longer has the right of withdrawal" : "le client n'a plus le droit de rétractation"),
-          (this.isFrench ?
-              "from the date and time of provision of services. By way of derogation from article" :
-              "à compter de la date et heure de fourniture de services. Par dérogation à l'article"
-          ),
-          (this.isFrench ?
-              "L. 121-20-1 of the French Consumer Code, the customer has a right of withdrawal of 7 (seven) days if the service has not yet been delivered." :
-              "L. 121-20-1 du Code Français de la consommation, le client a un droit de rétractation de 7 (sept) jours si le service n'a pas encore été livré."
-          ),
-          (this.isFrench ?
-              "This right of withdrawal is made by message to the email address" :
-              "Ce droit de rétractation s'effectue par message à l'adresse mail"
-          ) + ": contact@gaetandev.fr",
+          (isEnglish ? "Notwithstanding article L. 121-20-2 of the French Consumer Code," : "Par dérogation à l'article L. 121-20-2 du Code Français de la consommation,"),
+          (isEnglish ? "the customer no longer has the right of withdrawal" : "le client n'a plus le droit de rétractation"),
+          (isEnglish ? "from the date and time of provision of services. By way of derogation from article" : "à compter de la date et heure de fourniture de services. Par dérogation à l'article"),
+          (isEnglish ? "L. 121-20-1 of the French Consumer Code, the customer has a right of withdrawal of 7 (seven) days if the service has not yet been delivered." : "L. 121-20-1 du Code Français de la consommation, le client a un droit de rétractation de 7 (sept) jours si le service n'a pas encore été livré."),
+          (isEnglish ? "This right of withdrawal is made by message to the email address" : "Ce droit de rétractation s'effectue par message à l'adresse mail") + ": contact@gaetandev.fr",
         ],
-
-        "price": (this.isFrench ? "Prices" : "Tarifs"),
+        "price": (isEnglish ? "Prices" : "Tarifs"),
         "priceText": [
-          (this.isFrench ?
-              "The rates are communicated to the customer through a quote and an invoice." :
-              "Les tarifs sont communiqués au client grâce à un devis et une facture."
-          ),
-          (this.isFrench ?
-              "The gaetandev.fr site reserves the right to modify its prices at any time." :
-              "Le site gaetandev.fr se réserve le droit de modifier ses prix à tout moment."
-          ),
+          (isEnglish ? "The rates are communicated to the customer through a quote and an invoice." : "Les tarifs sont communiqués au client grâce à un devis et une facture."),
+          (isEnglish ? "The gaetandev.fr site reserves the right to modify its prices at any time." : "Le site gaetandev.fr se réserve le droit de modifier ses prix à tout moment."),
         ],
       }
     };
-
-    this.updateEvent.emit(this.isFrench);
   }
 }
